requirements.txt
flask
scapy
pillow 
pystray
werkzeug
psutil
flask-socketio
eventlet
schedule
reportlab
openpyxl
python-docx